resolved coments
