+++ sdk/cosmosdb/cosmos/src/client/ChangeFeed/buildChangeFeedIterator.ts
@@ -1,0 +1,158 @@
+// Copyright (c) Microsoft Corporation.
+// Licensed under the MIT License.
+import type { ClientContext } from "../../ClientContext";
+import type { PartitionKey } from "../../documents";
+import type { PartitionKeyRangeCache } from "../../routing";
+import { QueryRange } from "../../routing";
+import type { ChangeFeedIteratorOptions } from "./ChangeFeedIteratorOptions";
+import { ChangeFeedStartFrom } from "./ChangeFeedStartFrom";
+import { ChangeFeedStartFromBeginning } from "./ChangeFeedStartFromBeginning";
+import { ChangeFeedStartFromContinuation } from "./ChangeFeedStartFromContinuation";
+import { ChangeFeedStartFromNow } from "./ChangeFeedStartFromNow";
+import { ChangeFeedStartFromTime } from "./ChangeFeedStartFromTime";
+import { ChangeFeedResourceType } from "./ChangeFeedEnums";
+import { ChangeFeedForPartitionKey } from "./ChangeFeedForPartitionKey";
+import { ErrorResponse } from "../../request";
+import { ChangeFeedForEpkRange } from "./ChangeFeedForEpkRange";
+import { getIdFromLink, getPathFromLink, ResourceType, Constants } from "../../common";
+import {
+  buildInternalChangeFeedOptions,
+  fetchStartTime,
+  isEpkRange,
+  getEPKRangeForPrefixPartitionKey,
+} from "./changeFeedUtils";
+import { isPrefixPartitionKey, isPartitionKey } from "../../utils/typeChecks";
+import type { Container } from "../Container";
+import type { FeedRangeInternal } from "./FeedRange";
+import { PartitionKeyInternal } from "../../documents/PartitionKeyInternal";
+
+export async function buildChangeFeedIterator(
+  cfOptions: ChangeFeedIteratorOptions,
+  clientContext: ClientContext,
+  container: Container,
+  partitionKeyRangeCache: PartitionKeyRangeCache,
+): Promise<any> {
+  const url = container.url;
+  const path = getPathFromLink(url, ResourceType.item);
+  const id = getIdFromLink(url);
+
+  let changeFeedStartFrom = cfOptions.changeFeedStartFrom;
+
+  if (changeFeedStartFrom === undefined) {
+    changeFeedStartFrom = ChangeFeedStartFrom.Now();
+  }
+
+  if (changeFeedStartFrom instanceof ChangeFeedStartFromContinuation) {
+    const continuationToken = changeFeedStartFrom.getCfResourceJson();
+    const resourceType = changeFeedStartFrom.getResourceType();
+    const internalCfOptions = buildInternalChangeFeedOptions(
+      cfOptions,
+      changeFeedStartFrom.getCfResource(),
+    );
+
+    if (
+      resourceType === ChangeFeedResourceType.PartitionKey &&
+      isPartitionKey(continuationToken.partitionKey)
+    ) {
+      return new ChangeFeedForPartitionKey(
+        clientContext,
+        container,
+        id,
+        path,
+        continuationToken.partitionKey,
+        internalCfOptions,
+      );
+    } else if (resourceType === ChangeFeedResourceType.FeedRange) {
+      return new ChangeFeedForEpkRange(
+        clientContext,
+        container,
+        partitionKeyRangeCache,
+        id,
+        path,
+        url,
+        internalCfOptions,
+        undefined,
+      );
+    } else {
+      throw new ErrorResponse("Invalid continuation token.");
+    }
+  } else if (
+    changeFeedStartFrom instanceof ChangeFeedStartFromNow ||
+    changeFeedStartFrom instanceof ChangeFeedStartFromTime ||
+    changeFeedStartFrom instanceof ChangeFeedStartFromBeginning
+  ) {
+    const startFromNow = changeFeedStartFrom instanceof ChangeFeedStartFromNow ? true : false;
+    const startTime = startFromNow ? undefined : fetchStartTime(changeFeedStartFrom);
+
+    const internalCfOptions = buildInternalChangeFeedOptions(
+      cfOptions,
+      undefined,
+      startTime,
+      startFromNow,
+    );
+    const cfResource = changeFeedStartFrom.getCfResource();
+    if (isPartitionKey(cfResource)) {
+      const partitionKey = cfResource as PartitionKey;
+      const partitionKeyDefinition = await container.getPartitionKeyDefinition();
+
+      if (
+        partitionKeyDefinition !== undefined &&
+        isPrefixPartitionKey(partitionKey, partitionKeyDefinition.resource)
+      ) {
+        const effectiveEPKRange = await getEPKRangeForPrefixPartitionKey(
+          partitionKey as PartitionKeyInternal,
+        );
+        return new ChangeFeedForEpkRange(
+          clientContext,
+          container,
+          partitionKeyRangeCache,
+          id,
+          path,
+          url,
+          internalCfOptions,
+          effectiveEPKRange,
+        );
+      }
+
+      return new ChangeFeedForPartitionKey(
+        clientContext,
+        container,
+        id,
+        path,
+        cfResource as PartitionKey,
+        internalCfOptions,
+      );
+    } else {
+      let internalCfResource: QueryRange;
+      if (cfResource === undefined) {
+        internalCfResource = new QueryRange(
+          Constants.EffectivePartitionKeyConstants.MinimumInclusiveEffectivePartitionKey,
+          Constants.EffectivePartitionKeyConstants.MaximumExclusiveEffectivePartitionKey,
+          true,
+          false,
+        );
+      } else if (isEpkRange(cfResource)) {
+        internalCfResource = new QueryRange(
+          (cfResource as FeedRangeInternal).minInclusive,
+          (cfResource as FeedRangeInternal).maxExclusive,
+          true,
+          false,
+        );
+      } else {
+        throw new ErrorResponse("Invalid feed range.");
+      }
+      return new ChangeFeedForEpkRange(
+        clientContext,
+        container,
+        partitionKeyRangeCache,
+        id,
+        path,
+        url,
+        internalCfOptions,
+        internalCfResource,
+      );
+    }
+  } else {
+    throw new ErrorResponse("Invalid change feed start location.");
+  }
+}
